+++ src/main/java/com/wetwater/washwater/mixin/MixinLevelChunkSection.java
@@ -5,9 +5,6 @@ import com.wetwater.washwater.FluidSection;
 import com.wetwater.washwater.FluidSectionContainer;
 import com.wetwater.washwater.WaterInfo;
 import com.wetwater.washwater.state.WashWaterFluidState;
-import me.jellysquid.mods.lithium.common.block.BlockCountingSection;
-import me.jellysquid.mods.lithium.common.block.TrackedBlockStatePredicate;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.chunk.LevelChunkSection;
 import net.minecraft.world.level.material.FluidState;
@@ -16,8 +13,6 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import java.util.function.Predicate;
 
 @Mixin(value = LevelChunkSection.class, priority = 1200)
 public class MixinLevelChunkSection implements FluidSectionContainer {
@@ -58,60 +53,6 @@ public class MixinLevelChunkSection implements FluidSectionContainer {
             cir.setReturnValue(((FluidState)(Object)waterState));
         }
     }
-
-    //LITHIUM COMPAT
-
-/*    @Inject(at = @At("RETURN"), method = "maybeHas", cancellable = true)
-    public void maybeHas(Predicate<BlockState> predicate, CallbackInfoReturnable<Boolean> cir) {
-        //System.out.println("Test1");
-        if (fluidSection != null && !cir.getReturnValue()) {
-            //System.out.println("Test2");
-            if(predicate.test(Blocks.WATER.defaultBlockState())) {
-                //System.out.println("Test3");
-                cir.setReturnValue(!fluidSection.isEmpty());
-            }
-        }
-    }*/
-
-
-/*    public boolean anyMatch(TrackedBlockStatePredicate predicate, boolean cir) {
-        if (fluidSection != null && !cir) {
-            if(predicate.test(Blocks.WATER.defaultBlockState())) {
-                System.out.println("E");
-                cir = (!fluidSection.isEmpty());
-            }
-        }
-        return cir;
-    }*/
-/*    @Override
-    public boolean anyMatch(TrackedBlockStatePredicate predicate, boolean cir) {
-        if (fluidSection != null && !cir) {
-            if(predicate.test(Blocks.WATER.defaultBlockState())) {
-                cir = (!fluidSection.isEmpty());
-            }
-        }
-        return cir;
-    }*/
-
-    // END
-
-    /**
-     * @author
-     * @reason
-     */
-/*    @Overwrite
-    public boolean maybeHas(Predicate<BlockState> predicate) {
-        System.out.println("Test1");
-        if (fluidSection != null) {
-            System.out.println("Test2");
-            if (predicate.test(Blocks.WATER.defaultBlockState())) {
-                System.out.println("Test3");
-                return !fluidSection.isEmpty();
-            }
-        }
-        return false;
-    }*/
-
 
     @ModifyReturnValue(at = @At("RETURN"), method = "hasOnlyAir")
     public boolean hasOnlyAir(boolean original) {
