+++ src/main/java/com/wetwater/washwater/mixin/MixinLevelChunkSectionLithium.java
@@ -16,8 +16,6 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import java.util.function.Predicate;
-
 @Mixin(value = LevelChunkSection.class, priority = 1200)
 public class MixinLevelChunkSectionLithium implements FluidSectionContainer {
     @Unique
